Fixed bug 28482: Exchange rate and Converted Amt in Add Payment
The Exchange rate and Converted amount fields were not properly updated when 2 financial accounts (the first one of them with different currency than the invoice and the second one with the same currency) are selected sequentially into the Add Payment process.

Casue: the conversion rate was wrongly sent back to the JS as a String ("1") instead of using a number. The OB.APRM.AddPayment.paymentMethodMulticurrency function checks the received conversion rate is a number before updating the converted amount. Since the converted amount wasn't a number but a string, the amounts were not updated.
After the fix, the conversion rate is sent back using a integer 1

--HG--
extra : rebase_source : 58f597056b3181d1eb2058b7375f0e1e1af3850a