Fixed issue 28466: Autosave was executed when creating a new record.
When a window only has one subtab, and that subtab has a display logic, just after clicking the proper button to create a new
record in the header tab (both in grid and in form mode), the autosave is executed. This happens only the first time that a record is
created.

The problem was in the "updateSubtabVisibility" function of the "ob-standard-view.js".

The problem was that at first, in the function, there was a logic that was not taking into account the fact that if it was the record selected in the header tab has not been saved yet, its subtabs are empty, and it is not neccesary to refresh them.

Before the fix it tried to refresh the subtabs, which resulted in a call to autosave.

For fixing it a new condition has been added to the "updateSubtabVisibility" function. With this new condition
now, the call to the autosave is avoided in the case of the new record.

A new function called "isEditingNewRecord" has been created too. This function returns true if a new record is being edited in the view.