Fixes bug 28396:Focus not placed in field group if form has no editable fields
The issue 28396 [1] describes a problem that happens when the first focused field is a field group. This other issue [2] describes a very similar problem, the only difference beinge that instead of being reproduced in general field groups, it was reproduced for the specific audit field group. Its fixes [3], [4] only solved the problem for that particular field group.

The present changeset fixes the problem in a more general way. Instead on checking the type of the field being focused (that will change depending of whether the field group is general or specific (audit, notes, attachments, etc) it calls isc.isA.SectionItem to see whether the focused field is a field group.

[1] https://issues.openbravo.com/view.php?id=28396
[2] https://issues.openbravo.com/view.php?id=28231
[3] https://code.openbravo.com/erp/devel/pi/rev/a72704b89964c9dd369c422ccefeecca16e26551
[4] https://code.openbravo.com/erp/devel/pi/rev/c523abe0e8115d406c7e2515be7aef9d5209ce73

--HG--
extra : rebase_source : 1b6c7cc60707e9ec743cc0494cc0e032de94ecc6