Fixes issue 24577: Direct link works properly in grids with implicit filters
When a view was open by pasting a URL link (either in the address bar or in the quick view), the implicit filter of the header grid was always applied, even if the filters had been clea
red before obtaining the link.

This has been fixed by storing in the URL info about the grid filter clause state when the link is obtained. Then, when opening a view from that link, if there were no filter clauses wh
en the link was obtained, the grid will be initialized without them.

--HG--
extra : rebase_source : 0e9227980429c00ff5f9b8528dd58d03c7c26341