Fixes bug 28502: Inv Amt Update takes into account Costing Rule warehouse dim
The problem was that the process was checking if the warehouse in Inventory Amount Update is null or not, and if not, it uses it. This is correct when the Costing Rule is set as Warehouse Dimension, but if is not, it should not use it, and even if the Costing Rule is not set as Warehouse Dimension and the warehouse field is hidden in Inventory Amount Lines, it is filled with the Warehouse set in session value.

--HG--
extra : rebase_source : cd5794969c6729dd627789bf0eb89c57279b35ec