Fixes Bug 28487: Fixed Null Pointer Exception.
The problem was that in updateBDCostingTimeRange function inside AverageCostAdjustment class, the code was trying to get the movement date of a Inventory Transaction related to a Costing record. But not all costing have Inventory Transaction, so it was giving a NPE.
Now, if there is a previous Costing record than the one created due to a backdated transaction, this Costing will be shortened till transaction process date, and the new Backdated costing record will start at that moment.
This only happens the first time that a related transaction is found that is not a backdated transaction, because if is done with a backdated transaction time overlap can happen.

--HG--
extra : transplant_source : %21a%BD%F9%12%25%5E%B7UK%0Ei%14%18rz%1C%CE%EE%A5