Fixes issue 28505: Document number field length is not always hardcoded to 20
In this changeset [1], that was included in the first release of Openbravo 3.0, the length of the document number field length for the grids is hardcoded to 20. The length of the form field is taken from the length of the column in the application dictionary.

Having a different length in the grid and form fields resulted in the value of the field being truncated when it is edited in the view with the shortest length. It is not possible to remove the hardcoded code because it is possible that some Openbravo instances depend on it (i.e. if the length of the document number column is set to 0). In order to fix this issue, now the hardcoded code only affects the fields whose length is 0. For all the other cases, the length of the grid field will be taken from the length of the column.

[1] https://code.openbravo.com/erp/devel/pi/rev/ae5c327c567f#l3.14