Fixed issue 28500: Add targetEntity property the OBSelectorItems.
The "targetEntity" property didn't exist for the OBSelectorItems included in Parameter Windows.

The problem was that in the "ob-param-window-parameter.ftl" the logic for the "targetEntity" property
didn't exist.

For fixing this the proper logic has been added to the "ob-param-window-parameter.ftl" and a new method
called "getTargetEntity" has been performed in the "OBViewParameterHandler.java" class.
This method returns the "targetEntity" name of an OBSelector item which is included in a Parameter Window.