Fixed bug 28325: Journal Entries Detail fails depending on the date format.
When retrieving Fact Acct Group information in ReportGeneralLedgerJournalDetail class in a direct command request, now is taken into account that the date format defined could be dd/MM/yyyy, instead of default one dd-MM-yyyy, so the called of split function getValue, was changed to separate the string parameter by the position that realy marks the end of the date and the begins of the Group Id.
Now the complete date is read no matter the format defined, and the Journal Entries Detail form shows the information successfuly.

--HG--
extra : rebase_source : 3d52312b951159072322de7e01b6af279891147d