Fixed bug 28012: Sales dimensional report is not showing some dimensions.
New dimensions like project and ship to address were added to the Sales dimensional report to make possible to also organize data by them.
Now it is possible to also include project and ship to address dimensions when the report data is retrieved both in its version comparative and non-comparative.