Fixes bug 28554: No error is shown now in oracle creating LC from Goods receipt
The problem was in 'SELECT PROCESSED INTO v_Processed FROM M_LandedCost WHERE M_landedcost_ID=v_M_LandedCost_ID' statement. If no result is given in the select, in postgres no value is inserted in v_Processed but it works, but in oracle it fails. To fix this an check has been added to execute this statement only when v_M_LandedCost_ID is not null.
An if statement has been added in case v_M_LandedCost_ID is null and v_m_Inout_ID is not (Goods Receipt window), to prevent adding, updating or inserting Landed Cost Cost records when the Goods Receipt is processed.
Finally, in case of updating, the restriction of not being able to change landed cost header has been removed. This was causing problems when launching Cost Background for a Goods Receipt with Landed Cost Cost records, because this process creates the Landed Cost Header and assigns to Landed Cost Cost records.