Fixed issue 28597: Documents are not synced when quotation is loaded
Added the attribute 'isEditable' with value 'false' to the quotation at the point of closing it. This has been done to prevent discounts from executing when loading the quotation.
The problem was that, with the old discounts flow, the discounts calculation was saving the quotation in local database incorrectly.

--HG--
extra : rebase_source : 8ab294b06ed86cfe05e8df509ab29c5c954d4e03