Fixed bug 28452: Wrong Trial Balance having standard accounting in Opening date.
The method calculateTree, in Report Trial Balance process class, was calculating in a wrong way the balance of accounting's credits and debits in opening dates. So the method was changed to not just overwrite credit and debit values but to increase them with the accumulate, when initial balance information is requested.
Now when executing Trial Balance report, for opening dates, accounting's credits and debits are calculated properly and balance is set to 0.