Fixed issue 28623: Window's behaviour changes when creating a new record.
When a window is opened using the "quick create new" widget, the same window in the recent views layout
does not allow to create a new record (the green small button for creating a new record, which should be
by the window name, disappears).

The problem was in the "ob-utilities.js", inside the "{OB.Utilities.openView}" function, the logic for
setting the "singleRecord" value (this is the Edit Only attribute) was wrong. If the value of the
variable singleRecord of that function was undefined or null, as it was in the case of clicking the "quick create new"
widget, that value was set to true and the window became an "edit only" window.That was why the small
green "create new" icon disappeared from the recent views layout.

For fixing this the condition has been removed, so we know that the value of the variable "singleRecord" is not
going to be modified and it will be set correctly.

--HG--
extra : transplant_source : %E3%B4F%F2%E3%B0%07%7FO%B3%3A%E6%1E%9F%B5%F6%1B-%5E%B5