Fixed issue 28394: Cash up creation date is not properly set in the ERP when the cash up is synchronized
Changed the format of the creationDate of the cashup in the POS to match the format of the date when it's recovered from the server.
Added the createdDate field to the cashup json sent to the backend to be synchronized.
Audit info is updated in the server after the Dal object is saved for the first time.

--HG--
extra : rebase_source : 1b34254a1993f6a43038fc95d048dc61d53b316c