Fixes bug 28640:Res. line with amt released not duplicated in Manage Reservation
The problem is that when opening Manage Reservations P&E from sales order line, the property 'reservationStock' comes blank, even when a reservation stock line exists (When there is no reservation for the sales orderline a temporary one is created just to display lines in the selector selected correctly, and after that is deleted. This happens in the datasource of the P&E). In this cases the Manage Reservations action handler deletes these lines y they exists and creates new ones. But this is not possible for already released lines, so it was creating a duplicate of those.

In order to fix this, the property 'id' is used (allways comes with a value) in the case that the reservation is released, because in this case for sure that the reservation stock line exists, so it will update the existing one instead of creating a new one.