Fixes issue 28427: Numbers are always rounded on blur
The blur function of the OBNumberItems only did the rounding when the getValue function of the item returned a string. This worked in most of the cases (i.e. standard grids and forms, d
ue to the edit form of the grid having a setTextualValue function, see [1]), but not in the parameter windows. In those windows, both numeric parameters and numeric columns in grid para
meters did not round the entered numbers before saving them.

To fix this, now the numbers are rounded on blur even if the value entered is already a number. To found the number these two functions are invoked:

  value = OB.Utilities.Number.JSToOBMasked(this.getValue(), this.typeInstance.maskNumeric, this.typeInstance.decSeparator, this.typeInstance.groupSeparator);
  this.setValue(OB.Utilities.Number.OBMaskedToJS(value,  this.typeInstance.decSeparator, this.typeInstance.groupSeparator));

The first one converts the number to a string representation of the rounded number. The second one converts the rounded string number to a javascript number.

--HG--
extra : rebase_source : efd276c4425cf80c96013026a426909e5d89e7fd