Fixes bug 28747,related with bug 28427: Number entered with formula is rounded
The problem was that if a number was entered using a formula (i.e. by entering =1/3 in the form item), the value was not rounded using the numeric field mask. For non-formula inputs the validation is done at the beginning of the blur function. For formula inputs it is not possible to do it at that point, because the value of the field is still the formula, not the result of evaluating the formula. The formula is evaluated in the validate function(), so the rounding for formula inputs is done after invoking it.

--HG--
extra : transplant_source : 9%B9%17%1D%E0%17%AF%A0%0BC1%27%5C%11%1Df%25%1CL%B4