Fixes bug 28814: NPE fixed in Costing Background for a production product
When a product is defined as production the costing records created in the system are always for * organization. If a transaction of this product it was creating a backdated cost adjustment, in one place of the code a NPE exception error was happening because a costing record of legal entity org it was being trying to get. This has been fixed always filtering by client instead of organization when the product is of production type.

--HG--
extra : transplant_source : %CB%A02%BA%FC%B0%A9%18%C1%CF%F7d%B2m%05%F9%C5%C1%CE%11