Fixes issue 28747: Numeric field textual value is also rounded
The problem was that even though the value of the field was being properly rounded, its textual value - sent as the propertyName_textualValue - was being kept unrounded. To fix it, now when the value of a numeric field is entered using a formula both its numeric and its textual value are updated.