Fixes issue 28736: Selected record in parent tab criteria is applied to trees
In order to show the proper records in the subtabs, a criteria is applied based on the record selected in the parent tab. This criteria was automatically being added in the trees based
on the ADTree structure, but not in those based on the LinkToParent structure.

Now this criteria is also added in the LinkToParent trees if the following conditions are met:
- The tab is not a header tab
- The tree does not define a HQL clause for root nodes
- The parameters contains the info needed to build the new criteria

The getParentRecordIdFromCriteria method, which was used in the ADTreeDatasourceService for this purpose, has been moved up to TreeDatasourceService so that it can be reused by LinkToPa
rentTreeDatasourceService.

--HG--
extra : rebase_source : eaa070712da5b69f72216d27a6f21c1ac549c34a