Fixed bug 28691: It is possible to add discounts to documents already processed.
One possible solution for this issue would be to configure the C_ORDERDISCOUNT_TRG to be executed ON INSERT action, but this trigger is launched also  when inserting orders for business partners with discounts defined, so the trigger logic causes a "mutating table" oracle's error.
In this case a new event handler is developed, to launch an error message, when trying to insert discounts for a sales order already processed. This change also affects the following windows: Purchase Order, Return from Customer, Return to Vendor and Sales Quotation.