Related with issue 28736: Adds javadoc and final modifier to method
The getParentRecordIdFromCriteria of the TreeDatasourceService is not meant to be overwritten, so the final modifier has been added. The method cannot be made private because it is used by two of the TreeDatasourceService subclasses.

Also, javadoc for this method has been added.

--HG--
extra : rebase_source : fbf6d187b6c4c3b0ff3eb433913b61a75b49bd58