Fixes bug 28852: Adding automatically a tree node does not trigger error in log
In this changeset [1], a flush operation was added after the add, update and delete operations on ADTreeNode entities. This was done because it was needed for the update operation, and was also included fo
r the add and delete operations because it was thought that it would be needed too. The problem was that not only it was not needed, but it also shows an error in the log (the ADTreeNodes are succesfully a
dded/removed, though).

In this case, when an ADTreeNode is updated (i.e. when is moved under another node), the flush done to actually commit the changes must be done just after modifying it to ensure it is done while the contex
t is in admin mode. Otherwise, the next flush that is done might be outside an admin mode block, and the operation will fail if the user had no access to the ADTtreeNode entity. This does not happen with t
he add and remove operation. As long as the OBDal.getInstance().add and OBDal.getInstance().remove methods are invoked in an admin mode block, the operation will work regardless of when the OBDal.getInstan
ce().flush method is invoked.

[1] https://code.openbravo.com/erp/devel/pi/rev/5cec8a63b6c8

--HG--
extra : rebase_source : 319e2cc4474176f4db84cda5a0ae135d41471520