Fixes issue 28843: Characteristic selector does not filter by the selected item
The problem was that if a value was selected in a form item that uses the tree references, when the user opened the drop down the values were filtered using the selected value. This is not the correct behaviour, if the user opens manually the dropdown all the values that comply with the where clause should be shown.

This happened because each time the picklist was shown, its values were retrieved filtering the selected value. Now if the user is manually opening the tree item, the result records will not be filtered by the selected record.

Also a couple of changes have been done in the getCriteriaFromTreeItem function:
- For the criteria the entered value (the text currently shown in the tree item) is used, instead of the selected value.
- A not-null check has been added for the valueMap of the item before trying to use it.

--HG--
extra : rebase_source : ade056b0e420b17e20c46306ffa3a6630e095abb