Fixes bug 28872: Tree item popup is loaded properly when using a partial filter
The problem was that if a partial filter was entered in the tree item (as opposed to a tree item value selected), when the popup was opened that partial filter was being used to filter the nodes by their id.

Now before creating that criteria it is checked that the value entered in the tree item actually is the id of a selected item.