Fixed issue 28628: The header and footer components of the ticket should be extensible
Added an OrderFooter component to be able to add components to the bottom of the ticket.
The OrderFooter component includes an array to contain new components.
Refactored the OrderHeader to include two arrays of components, one for labels and the second for buttons. These arrays can be extended from other modules.