Fixes bug 28111, Fixes bug 28951: InitCostDateacct modulescript has been removed
The functionality that was being implemented in the modulescript now will be done in Costing Background process. An oncreatedefault has been added in dateacct column because it has a constraint of not null. The date set to the column will be 01-01-1970, and this date will be used as a reference while updating it in Costing Background process