Fixes Bug 28962: Cost properly calculated for orphan lines in Cost Adjustments
The process was failing for orphan lines because it was trying to get the cost from a Purchase Price or an Standard Cost, and there is none. Before checking those two, the process now tries to find an Average Cost for that product on the movement date of the orphan line, and if it founds it assigns to the transaction.