Fixed issue 28761: A tab will be empty when the user role has readOnly access
The problem was that if a user role has readOnly access to a particular tab and this tab is
opened by using "open on tab", an empty tab is displayed. This problem was because there was
a invalidate cache between a request to the datasource and receipt of the answer.


The cache invalidation occurred because refreshContents() were called in "setReadOnlyMode".
This cache invalidation generated a bad behaviour for displaying the tab.

To check this, It has been obtained a trace of the functions that are invoked. Briefly, after
each call to the DataSource it must be seen as the data arrives (arrives in fetchRemoteDataReply())
 without invalidating cache. It should be noted, in the first call to DataSource we can find an
invalidateCache (main problem):

*** 2 *** Run invalidateCache() for isc_OBViewGrid_7. (ResultSet.js)
*** 1 *** Call to DataSource: sendDSRequest for isc_OBViewGrid_7. (DataSource.js)
*** 2 *** Run invalidateCache() for isc_OBViewGrid_7. (ResultSet.js)
*** 1 *** Call to DataSource: sendDSRequest for isc_OBViewGrid_7. (DataSource.js)
*** 3 *** Run fetchRemoteDataReply() for isc_OBViewGrid_7. (ResultSet.js).
It has been detected that the cache is invalidated. (ResultSet.js)
*** 3 *** Run fetchRemoteDataReply() for isc_OBViewGrid_7. (ResultSet.js)

To avoid this problem, it has been deleted method refresContent() in method setReadOnlyMode() because
it was not necessary. This action avoids the invalidation of the cache.