Fixes issue 28985: Grid columns can be frozen
In this changeset [1], a function named fieldIsVisible was added to the OBViewGrid class, to control if a field was actually visible in the grid. The problem was that there was actually a function defined with that same name in the Smartclient's DataBoundComponent class, which is an OBViewGrid ancestor. This means that the fieldIsVisible function implemented in [1] is overwriting the original implementation, and this was not intended.

The fieldIsVisible function is invoked when a grid column is frozen, that is why the change made in [1] has created the current issue. To fix it, the name of OBViewGrid.fieldIsVisible function has been renamed to fieldIsVisibleInGrid. It has been checked that there is no function defined with that name in any OBViewGrid ancestor class.

[1] https://code.openbravo.com/erp/devel/pi/rev/25794

--HG--
extra : transplant_source : %AB%C5iz%08r%E73%AE%85F%A3%AD-k%23%00l%ED7