Fixed bug 28606: Tax Amount field of Invoice Lines should be deprecated
Tax Amount field of Invoice Lines has been deprecated as it is calculated only
when the lines are created manually and if the tax rate is a summary tax rate
the value calculated is not correct. The field is never displayed.

Now the field Development Status from the column taxAmt (c_invoiceline table)
has been changed from 'Ready' to 'Deprecated'.