Fixes bug 29030: Match Statement does not fail when a line throws an exception
In Matching Algorithm in Match Statement, if a line to be matched fails for any reason (period closed for example) a rollback occurs. This rollback was not being properly managed in Matching Algorithm after the refactor of Match Statement window, so the process was failing instead of skipping that line and going forward. Now it skips and the process finished successfully.

--HG--
extra : transplant_source : %A1i%E4i9%DB%B7%40%C3%85%BEj%20%27%91%FB%E7%0E%3C%9E