Fixes bug 29032: Arguments to the selectionChanged function are passed properly
Before this changeset [1] was committed, the OBPickAndExecuteGrid.selectionUpdated(record, selected) function did not take into account the second parameter. This parameter was set to true if the record was being selected and to false if it was being deselected.

Now that the selected parameter is being taking into account, some APRM logic that should be executed when a record in a P&E window is selected is not being run. The problem was that tje selection updated function was being called incorrectly, like this:

this.Super('selectionChanged', record, state);

When invoking the Super function, all the parameters that are to be passed to the function should be included in the second argument (as an array if there are more than one). This is the way it should be done:

this.Super('selectionChanged', [record, state]);

[1] https://code.openbravo.com/erp/devel/pi/rev/d061c9e837aa5a0ce0a7fb221e1ac8c822748988