Fixes bug 28987: Add payment does not take long to load with lot of credit paym.
When opening Add Payment process definition it has to load the default values of the parameters of the process definition. One of this parameters is the credit of the business partner of the payment. The method that calculates this credit calls the method seqnumberpaymentstatus that uses CallStoredProcedure.getInstance().call to call the database function. This method is executed for each credit payment of the customer and on each execution an unnecessary flush was being done, so with lots of credit payments the impact on performance was huge.
A false flag has been added to call process to not to do flush on each execution of the process, so the performance problem is avoided.

--HG--
extra : rebase_source : db719db6a607f3e33e68bb28432f63ff8e217ecf