Fixed bug 28948: Orders Awaiting Delivery report do not show well the Stock Qtty
Orders Awaiting Delivery report's query, was not obtaining correctly the Quantity in Stock, due to a wrong comparison with the value of the attribute set of products in order's line, when it is seted as null.
To avoid this, the report's sql query was changed to compare the attribute set values of the storage detail between them, if the order's lines product has not an attribute set value defined. In other case, are compared the attribute set values of the order's line and the storage detail.
Now the product's stock quantity is correctly shown.