Fixes issue 29010: Grid is actually filtered using relative dates
The problem was that when a relative date filter was added to the grid and then used in a saved view, the saved view will not apply the relative date filter, but an absolute one. This was caused by this code in the OBViewGrid.convertCriteria function:

    if (this.dataSource) {
      criteria = this.dataSource.convertRelativeDates(criteria);
    }

Every time that a criteria is sent to the backend (i.e. when storing a saved view), the criteria objects that contained relative dates wwere converted to absolute dates. This was done to fix this issue
 [1], its problem being that the datasource used to populate the filter drop down did not support relative dates.

The function where that code was placed was too central, the criterias converted should have been only the ones used to populate the filter drop downs, but all the criterias were being converted. To fix this, the code has been adapted and moved to the OBFKFilterTextItem.getPickListFilterCriteria function.
[1] https://issues.openbravo.com/view.php?id=27679

--HG--
extra : transplant_source : %96E%A9%9F%136%3F%F7%1B%E9%A0%F1rOK%E5%85%3A%D9Z