Fixed issue 28857:Minimized view of query-list widgets sort data in the client
The problem was that in minimized view, the query-list widgets just sort the data
in the client. So it is only showing the data being displayed locally, instead of
requesting the data to the server.

This happens because "useClientSorting" was not setting properly. This is related
to the change of leaving of calculate total rows to improve performance.

Now, it works properly in minimized and maximized view because it has been setting
properly "useClientSorting" variable. The exceptional case is when the widget is in
minimized view and "ShowAllData" is not active. In this case, it will always make a
request for the data.