Fixes bug 29063: Fixed mutating trigger in Oracle.
M_RESERVATION_TRG was triggering a mutating trigger error because a select to the m_reservation table was being done on that trigger, that belongs to m_reservation table. That select has been modified to now to join m_reservation table, because it is not necessary.

--HG--
extra : rebase_source : 37bfda2e03b45adc77d5314790010c14e2ca34f7