28620: Button reference is not displayed properly in parameter windows
Button reference is not displayed properly in parameter windows, currently
in the User Interface a checkbox is drawn (like in the Yes/No reference).

Now, It is already possible to add new buttons to Process Definition by using
the "ButtonList" reference. So, the solution for this issue has been to remove
the "Button" reference from the list of references that can be assigned
for the parameters.