Fixes bug 29190: getStartingDate() method does not fail in Costing Background.
An OBQuery is being executed in getStartingDate() method inside AverageAlgorithm. There are 2 parameters in this OBQuery that are not setting properly, 'client' and 'org'. The whereclause is filtering by 'id' of these properties and instead of passing the id of these objects, the object was being passed to the OBQuery. These sometimes works but other times don't, so this has been changed to always setting the id's as parameters.

--HG--
extra : rebase_source : 84ef4d30402ddf70219fba67efa77bd0956a126a