Fixes bug 29079: Fixed Costing Background infinite loop.
Costing Background process, due to Cost Adjustments new implementation, on a certain point of the code, it was not applying correctly a setScale method to round a value of the expected cost. This was causing to have allways a difference between the expected cost and unitcost. Because of this, an adjustment was being created to adjust this difference, and to adjust all related transactions, causing an infinite loop

--HG--
extra : rebase_source : 02eb1d30945663dd62025bc7497b4518610d0d76