Fixed issue 28513: Error message wrongly shown when processing an inventory.
The error message when processing an inventory is wrongly shown with a product
with % in its name and a negative stock.

The issue occurs when we are parsing the error message and detects the % symbol
and try to find the string format in the next character.

The proposal solution is to eliminate the String.format() method from the
code.

Now, the error message is shown properly with the correct name.