Fixed bug 27010: Invoice Dimensional Report should not consider voided status.
A new primary filter, was added to the Sales Invoice Dimensional Report, to allow exclude or not, documents voided in report's output.
This filter can be managed through a new preference that was created too, in order to let user decide if filter will be shown. By default the filter will be hidden and documents voided will be considered.